backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.8
Django==5.1.15
django-celery-beat==2.7.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
kombu==5.6.2
packaging==26.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
